% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddo_ddf_methods.R
\name{as.list.ddo}
\alias{as.list.ddo}
\title{Turn 'ddo' / 'ddf' Object into a list}
\usage{
\method{as.list}{ddo}(x, ...)
}
\arguments{
\item{x}{a 'ddo' / 'ddf' object}

\item{\ldots}{additional arguments passed to \code{as.list}}
}
\description{
Turn 'ddo' / 'ddf' Object into a list
}

