% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recombine_combine.R
\name{combCollect}
\alias{combCollect}
\title{"Collect" Recombination}
\usage{
combCollect(...)
}
\arguments{
\item{\ldots}{...}
}
\description{
"Collect" recombination - simply collect the results into a local list of key-value pairs
}
\details{
This is an experimental prototype.  It is to be passed as the argument \code{combine} to \code{\link{recombine}}.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{divide}}, \code{\link{recombine}}, \code{\link{combDdo}}, \code{\link{combMeanCoef}}, \code{\link{combRbind}}, \code{\link{combMean}}
}

