% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombine_combine.R
\name{combDdf}
\alias{combDdf}
\title{"DDF" Recombination}
\usage{
combDdf(...)
}
\arguments{
\item{\ldots}{...}
}
\description{
"DDF" recombination - simply results into a "ddf" object, rbinding if necessary
}
\details{
This is an experimental prototype.  It is to be passed as the argument \code{combine} to \code{\link{recombine}}.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{divide}}, \code{\link{recombine}}, \code{\link{combCollect}}, \code{\link{combMeanCoef}}, \code{\link{combRbind}}, \code{\link{combDdo}}
}

