\name{combDiv}
\alias{combDiv}
\title{Division Recombination}
\usage{
  combDiv(...)
}
\arguments{
  \item{\ldots}{...}
}
\description{
  Division recombination (output of recombine is an object
  of class localDiv or rhDiv)
}
\details{
  This is an experimental prototype.  It is to be passed as
  the argument \code{combine} to \code{\link{recombine}}.
}
\author{
  Ryan Hafen
}
\seealso{
  \code{\link{divide}}, \code{\link{recombine}}
}

