% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recombine_combine.R
\name{combMean}
\alias{combMean}
\title{Mean Recombination}
\usage{
combMean(...)
}
\arguments{
\item{\ldots}{...}
}
\description{
Mean recombination
}
\details{
This is an experimental prototype.  It is to be passed as the argument \code{combine} to \code{\link{recombine}}.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{divide}}, \code{\link{recombine}}, \code{\link{combCollect}}, \code{\link{combDdo}}, \code{\link{combRbind}}, \code{\link{combMeanCoef}}
}

