% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombine_combine.R
\name{combMeanCoefNStdErr}
\alias{combMeanCoefNStdErr}
\title{Mean Coefficient Recombination and Standard Errors}
\usage{
combMeanCoefNStdErr(...)
}
\arguments{
\item{\ldots}{additional attributes to define the combiner (currently only used internally)}
}
\description{
Mean coefficient recombination -- Calculate the weighted average of
parameter estimates for a model fit to each subset
}
\details{
\code{combMeanCoefNStdErr} is passed to the argument \code{combine} in \code{\link{recombine}}

This method recombines the means exactly like \code{\link{combMeanCoef}}.
However, in additional to returning the recombined means of the
co-efficients, this function also returns the recombined Standard Errors.
As such, this function returns a list rather than a single value.
}
\author{
Hon Hwang
}
\seealso{
\code{\link{divide}}, \code{\link{recombine}}, \code{\link{rrDiv}}, \code{\link{combCollect}}, \code{\link{combDdo}}, \code{\link{combDdf}}, \code{\link{combRbind}}, \code{\link{combMean}}
}

