% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recombine_combine.R
\name{combRbind}
\alias{combRbind}
\title{"rbind" Recombination}
\usage{
combRbind(...)
}
\arguments{
\item{\ldots}{...}
}
\description{
"rbind" recombination
}
\details{
To combine divisions by row binding, this function can be passed as the \code{combine} argument of the
\code{\link{recombine}}.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{divide}}, \code{\link{recombine}}, \code{\link{combDdo}}, \code{\link{combCollect}}, \code{\link{combMeanCoef}}, \code{\link{combMean}}
}

