% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddo_ddf_methods.R
\docType{methods}
\name{ddf-accessors}
\alias{ddf-accessors}
\alias{splitRowDistn}
\alias{summary.ddo}
\alias{summary.ddf}
\alias{nrow}
\alias{NROW}
\alias{ncol}
\alias{NCOL}
\alias{nrow,ddf-method}
\alias{NROW,ddf-method}
\alias{ncol,ddf-method}
\alias{NCOL,ddf-method}
\alias{names.ddf}
\title{Accessor methods for 'ddf' objects}
\usage{
splitRowDistn(x)

\method{summary}{ddo}(object, ...)

\method{summary}{ddf}(object, ...)

nrow(x)

NROW(x)

ncol(x)

NCOL(x)

\S4method{nrow}{ddf}(x)

\S4method{NROW}{ddf}(x)

\S4method{ncol}{ddf}(x)

\S4method{NCOL}{ddf}(x)

\method{names}{ddf}(x)
}
\arguments{
\item{x}{a 'ddf' object}

\item{object}{a 'ddf'/'ddo' object}

\item{\ldots}{additional arguments}
}
\description{
Accessor methods for 'ddf' objects
}
\examples{
d <- divide(iris, by = "Species", update = TRUE)
nrow(d)
ncol(d)
length(d)
names(d)
summary(d)
getKeys(d)
}
