% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ddo_ddf_methods.R
\docType{methods}
\name{ddf-class}
\alias{NCOL}
\alias{NCOL,ddf-method}
\alias{NROW}
\alias{NROW,ddf-method}
\alias{ddf-class}
\alias{ncol}
\alias{ncol,ddf-method}
\alias{nrow}
\alias{nrow,ddf-method}
\title{'ddf' accessors}
\usage{
nrow(x)

NROW(x)

ncol(x)

NCOL(x)

\S4method{nrow}{ddf}(x)

\S4method{NROW}{ddf}(x)

\S4method{ncol}{ddf}(x)

\S4method{NCOL}{ddf}(x)
}
\arguments{
\item{x}{a 'ddf' object}
}
\description{
'ddf' accessors

'ddf' accessors

'ddf' accessors

'ddf' accessors

The Number of Rows/Columns of a 'ddf' object
}

