\name{ddf}
\alias{ddf}
\title{Make a distributed data frame}
\usage{
  ddf(conn, transFn = identity, update = FALSE,
    reset = FALSE, control = NULL, verbose = TRUE)
}
\arguments{
  \item{verbose}{logical - print messages about what is
  being done}
}
\description{
  Make a distributed data frame
}

