\name{setAttributes.ddf}
\alias{getAttribute}
\alias{getAttributes.ddf}
\alias{hasAttributes.ddf}
\alias{setAttributes.ddf}
\title{Managing attributes of 'ddo' or 'ddf' objects}
\usage{
\method{setAttributes}{ddf}(obj, attrs)

getAttribute(obj, attrName)

\method{getAttributes}{ddf}(obj, attrNames)

\method{hasAttributes}{ddf}(obj, attrNames)
}
\arguments{
  \item{attrs}{a named list of attributes to set}

  \item{obj}{'ddo' or 'ddf' object}

  \item{attrName}{name of the attribute to get}

  \item{attrNames}{vector of names of the attributes to
  get}
}
\description{
Managing attributes of 'ddo' or 'ddf' objects
}

