\name{divExampleKey}
\alias{divExampleKey}
\title{Get An Example Key of a Subset of Divided Data
Get the key for a subset of a divided data object to test functions on prior to doing computation across the entire data set}
\usage{
  divExampleKey(data)
}
\arguments{
  \item{divided}{data object (e.g. of class "localDiv",
  "rhDdata", etc.)}
}
\value{
  one subset of the divided data
}
\description{
  Get An Example Key of a Subset of Divided Data Get the
  key for a subset of a divided data object to test
  functions on prior to doing computation across the entire
  data set
}
\author{
  Ryan Hafen
}
\seealso{
  \code{\link{divide}}, \code{\link{rhipeDiv}},
  \code{\link{vdbPlot}}
}

