% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/divide.R
\name{divide-internals}
\alias{addSplitAttrs}
\alias{dfSplit}
\alias{divide-internals}
\title{Functions used in divide()}
\usage{
dfSplit(curDF, by, seed)

addSplitAttrs(curSplit, bsvFn, by, postTransFn = NULL)
}
\arguments{
\item{curDF,seed}{arguments}

\item{curSplit,bsvFn,by,postTransFn}{arguments}
}
\description{
Functions used in divide()
}
\note{
These functions can be ignored.  They are only exported to make their use in a distributed setting more convenient.
}

