\name{divide}
\alias{divide}
\title{Divide}
\usage{
  divide(data, ...)
}
\arguments{
  \item{data}{data to divide}

  \item{\ldots}{other arguments passed to the specific
  divide method}
}
\description{
  Divide data into subsets based on different criteria
}
\details{
  See the following for specific details for different data
  objects: \itemize{ \item{\code{\link{divide.data.frame}}:
  for data.frames} \item{\code{\link{divide.rhDF}}: for
  RHIPE data.frames} }
}
\author{
  Ryan Hafen
}

