% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/divide.R
\name{dfSplit}
\alias{addSplitAttrs}
\alias{dfSplit}
\title{Functions used in divide()}
\usage{
dfSplit(curDF, by, seed)

addSplitAttrs(curSplit, bsvFn, by, postTransFn = NULL)
}
\arguments{
\item{curDF,seed}{arguments}

\item{curSplit,bsvFn,by,postTransFn}{arguments}
}
\description{
Functions used in divide()
}

