\name{drApply}
\alias{drApply}
\title{Generic Recombination 'apply' Method}
\usage{
  drApply(fn, group = FALSE)
}
\arguments{
  \item{fn}{a function to apply to each subset}

  \item{group}{should results of function applied to each
  subset be grouped?}
}
\description{
  Generic recombination 'apply' method
}
\author{
  Ryan Hafen
}
\seealso{
  \code{\link{divide}}, \code{\link{recombine}}
}

