% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/globals.R
\name{drGetGlobals}
\alias{drGetGlobals}
\title{Get Global Variables and Package Dependencies}
\usage{
drGetGlobals(f)
}
\arguments{
\item{f}{function}
}
\value{
a list of variables (named by variable) and a vector of package names
}
\description{
Get global variables and package dependencies for a function
}
\details{
This traverses the parent environments of the supplied function and finds all global variables using \code{\link{findGlobals}} and retrieves their values.  All package function calls are also found and a list of required packages is also returned.
}
\examples{
a <- 1
f <- function(x) x + a
drGetGlobals(f)
}
\author{
Ryan Hafen
}

