% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataops_persist.R
\name{drPersist}
\alias{drPersist}
\title{Persist a Transformed 'ddo' or 'ddf' Object}
\usage{
drPersist(x, output = NULL, overwrite = FALSE, control = NULL)
}
\arguments{
\item{x}{an object of class 'ddo' or 'ddf'}

\item{output}{a "kvConnection" object indicating where the output data should reside (see \code{\link{localDiskConn}}, \code{\link{hdfsConn}}).  If \code{NULL} (default), output will be an in-memory "ddo" object.}

\item{overwrite}{logical; should existing output location be overwritten? (also can specify \code{overwrite = "backup"} to move the existing output to _bak)}

\item{control}{parameters specifying how the backend should handle things (most-likely parameters to \code{rhwatch} in RHIPE) - see \code{\link{rhipeControl}} and \code{\link{localDiskControl}}}
}
\value{
a 'ddo' or 'ddf' object
}
\description{
Persist a transformed 'ddo' or 'ddf' object
}
\examples{
bySpecies <- divide(iris, by = "Species")
bySpeciesSepal <- bySpecies \%>\%
 addTransform(function(x) x[,c("Sepal.Length", "Sepal.Width")]) \%>\%
 drPersist()
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{addTransform}}
}

