% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{flatten}
\alias{flatten}
\title{"Flatten" a ddf Subset}
\usage{
flatten(x)
}
\arguments{
\item{x}{a value of a key-value pair}
}
\description{
Add split variables and BSVs (if any) as columns to a subset of a ddf.
}
\examples{
d <- divide(iris, by = "Species")
# the column "Species" is no longer explicitly in the data
d[[1]]$value
# but it is preserved and can be added back in with flatten()
flatten(d[[1]]$value)
}
\seealso{
\code{\link{getSplitVars}}, \code{\link{getBsvs}}
}

