% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/divide.R
\name{getBsv}
\alias{getBsv}
\title{Get Between Subset Variable}
\usage{
getBsv(x, name)
}
\arguments{
\item{x}{a key-value pair or a value}

\item{name}{the name of the BSV to get}
}
\description{
For a given key-value pair, get a BSV variable value by name (if present)
}

