% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{getSplitVar}
\alias{getSplitVar}
\title{Extract "Split" Variable}
\usage{
getSplitVar(x, name)
}
\arguments{
\item{x}{a key-value pair or a value}

\item{name}{the name of the split variable to get}
}
\description{
For a given key-value pair or value, get a split variable value by name, if present (split variables are variables that define how the data was divided).
}

