\name{getSplitVars}
\alias{getSplitVars}
\title{Extract "Split" Variables}
\usage{
getSplitVars(x)
}
\arguments{
  \item{x}{a key-value pair or a value}
}
\description{
For a given k/v pair or value, exract all split variables
(split variables are variables that define how the data was
divided).
}

