% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvPairs.R
\name{kvPair}
\alias{kvPair}
\title{Specify a Key-Value Pair}
\usage{
kvPair(k, v)
}
\arguments{
\item{k}{key - any R object}

\item{v}{value - any R object}
}
\value{
a list of objects of class "kvPair"
}
\description{
Specify a key-value pair
}
\examples{
kvPair("name", "bob")
}
\seealso{
\code{\link{kvPairs}}
}

