\name{print.ddo}
\alias{print.ddo}
\title{Print a "ddo" or "ddf" Object}
\usage{
  \method{print}{ddo} (x)
}
\arguments{
  \item{x}{object to be printed}
}
\description{
  Print an overview of attributes of distributed data
  objects (ddo) or distributed data frames (ddf)
}
\author{
  Ryan Hafen
}

