% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvPairs.R
\name{print.kvPair}
\alias{print.kvPair}
\title{Print a key-value pair}
\usage{
\method{print}{kvPair}(x, ...)
}
\arguments{
\item{x}{object to be printed}

\item{\ldots}{additional arguments}
}
\description{
Print a key-value pair
}
\examples{
 kvPair(1, letters)
}

