\name{print.localDiv}
\alias{print.localDiv}
\title{Print a "localDiv" Object}
\usage{
  print.localDiv(x)
}
\arguments{
  \item{x}{object to be printed}
}
\description{
  Print a "localDiv" object
}
\details{
  Prints overview of attributes in objects of class
  "localDiv"
}
\author{
  Ryan Hafen
}

