\name{print.rhData}
\alias{print.rhData}
\title{Print an "rhData" or "rhDF" Object}
\usage{
  print.rhData(x)
}
\arguments{
  \item{x}{object to be printed}
}
\description{
  Print an "rhData" or "rhDF" object
}
\details{
  Prints overview of attributes in objects of class
  "rhData", "rhDF", as well as "rhDiv"
}
\author{
  Ryan Hafen
}

