\name{processTextFileByChunk}
\alias{processTextFileByChunk}
\title{Experimental sequential text reader helper function
Experimental helper function for reading text data in sequentially and adding to connection using \code{\link{addData}}}
\usage{
  processTextFileByChunk(inFile, outConn, chunkSize, fn,
    header = TRUE, skip = 0, recordEndRegex = NULL)
}
\description{
  Experimental sequential text reader helper function
  Experimental helper function for reading text data in
  sequentially and adding to connection using
  \code{\link{addData}}
}

