\name{rhText2df}
\alias{rhText2df}
\title{Experimental HDFS text reader helper function
Experimental helper function for reading text data into HDFS}
\usage{
  rhText2df(x, output = NULL, transFn = NULL, keyFn = NULL,
    linesPerBlock = 10000, control = NULL, update = FALSE)
}
\description{
  Experimental HDFS text reader helper function
  Experimental helper function for reading text data into
  HDFS
}

