% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapreduce_spark.R
\name{sparkControl}
\alias{sparkControl}
\title{Specify Control Parameters for Spark Job}
\usage{
sparkControl()
}
\description{
Specify control parameters for a Spark job.  See \code{rhwatch} for details about each of the parameters.
}

