\name{wMeanCoef}
\alias{wMeanCoef}
\title{Weighted Coefficient Mean Recombination}
\usage{
  wMeanCoef()
}
\description{
  Weighted coefficient mean recombination
}
\details{
  This is an experimental prototype.  It is to be passed as
  the argument \code{recomb} to \code{\link{recombine}}.
}
\author{
  Ryan Hafen
}
\seealso{
  \code{\link{divide}}, \code{\link{recombine}},
  \code{\link{rrDiv}}
}

