% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddo_ddf_methods.R
\name{as.data.frame.ddf}
\alias{as.data.frame.ddf}
\title{Turn 'ddf' Object into Data Frame}
\usage{
\method{as.data.frame}{ddf}(x, row.names = NULL, optional = FALSE,
  keys = TRUE, splitVars = TRUE, bsvs = FALSE, ...)
}
\arguments{
\item{x}{a 'ddf' object}

\item{row.names}{passed to \code{as.data.frame}}

\item{optional}{passed to \code{as.data.frame}}

\item{keys}{should the key be added as a variable in the resulting data frame? (if key is not a character, it will be replaced with a md5 hash)}

\item{splitVars}{should the values of the splitVars be added as variables in the resulting data frame?}

\item{bsvs}{should the values of bsvs be added as variables in the resulting data frame?}

\item{\ldots}{additional arguments passed to as.data.frame}
}
\description{
Rbind all the rows of a 'ddf' object into a single data frame
}
\examples{
d <- divide(iris, by = "Species")
as.data.frame(d)
}

