\name{combCollect}
\alias{combCollect}
\title{"Collect" Recombination}
\usage{
  combCollect(...)
}
\arguments{
  \item{\ldots}{...}
}
\description{
  "Collect" recombination
}
\details{
  This is an experimental prototype.  It is to be passed as
  the argument \code{combine} to \code{\link{recombine}}.
}
\author{
  Ryan Hafen
}
\seealso{
  \code{\link{divide}}, \code{\link{recombine}}
}

