\name{combDdo}
\alias{combDdo}
\title{"DDO" Recombination}
\usage{
combDdo(...)
}
\arguments{
  \item{\ldots}{...}
}
\description{
"DDO" recombination - simply collect the results into a
"ddo" object
}
\details{
This is an experimental prototype.  It is to be passed as
the argument \code{combine} to \code{\link{recombine}}.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{divide}}, \code{\link{recombine}},
\code{\link{combCollect}}, \code{\link{combMeanCoef}},
\code{\link{combRbind}}, \code{\link{combMean}}
}

