\name{combMeanCoef}
\alias{combMeanCoef}
\title{Mean Coefficient Recombination}
\usage{
  combMeanCoef(...)
}
\description{
  Mean coefficient recombination
}
\details{
  This is an experimental prototype.  It is to be passed as
  the argument \code{combine} to \code{\link{recombine}}.
}
\author{
  Ryan Hafen
}
\seealso{
  \code{\link{divide}}, \code{\link{recombine}},
  \code{\link{rrDiv}}
}

