\name{combMeanCoef}
\alias{combMeanCoef}
\title{Mean Coefficient Recombination}
\usage{
  combMeanCoef(...)
}
\arguments{
  \item{\ldots}{...}
}
\description{
  Mean coefficient recombination
}
\details{
  This is an experimental prototype.  It is to be passed as
  the argument \code{combine} to \code{\link{recombine}}.
  It expects to be dealing with named vectors including an
  element \code{n} specifying the number of rows in that
  subset.
}
\author{
  Ryan Hafen
}
\seealso{
  \code{\link{divide}}, \code{\link{recombine}},
  \code{\link{rrDiv}}
}

