% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recombine_combine.R
\name{combMeanCoef}
\alias{combMeanCoef}
\title{Mean Coefficient Recombination}
\usage{
combMeanCoef(...)
}
\arguments{
\item{\ldots}{...}
}
\description{
Mean coefficient recombination
}
\details{
This is an experimental prototype.  It is to be passed as the argument \code{combine} to \code{\link{recombine}}.  It expects to be dealing with named vectors including an element \code{n} specifying the number of rows in that subset.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{divide}}, \code{\link{recombine}}, \code{\link{rrDiv}}, \code{\link{combCollect}}, \code{\link{combDdo}}, \code{\link{combRbind}}, \code{\link{combMean}}
}

