% Generated by roxygen2 (4.0.1): do not edit by hand
\name{combRbind}
\alias{combRbind}
\title{"rbind" Recombination}
\usage{
combRbind(...)
}
\arguments{
\item{\ldots}{...}
}
\description{
"rbind" recombination
}
\details{
This is an experimental prototype.  It is to be passed as the argument \code{combine} to \code{\link{recombine}}.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{divide}}, \code{\link{recombine}}, \code{\link{combDdo}}, \code{\link{combCollect}}, \code{\link{combMeanCoef}}, \code{\link{combMean}}
}

