% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadr-package.R
\docType{package}
\name{datadr-package}
\alias{datadr}
\alias{datadr-package}
\title{datadr}
\description{
datadr: Divide and Recombine for Large, Complex Data
}
\details{
\tabular{ll}{
Package: \tab datadr\cr
Type: \tab Package\cr
Version: \tab 0.7.6.0\cr
Date: \tab 2015-10-19\cr
License: \tab BSD_3_clause + file LICENSE\cr
}
}
\examples{
help(package = datadr)
}
\author{
Ryan Hafen

Maintainer: Ryan Hafen <rhafen@gmail.com>
}
\keyword{package}

