% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadr-package.R
\docType{package}
\name{datadr-package}
\alias{datadr}
\alias{datadr-package}
\title{datadr}
\description{
datadr: Divide and Recombine for Large, Complex Data
}
\details{
\url{http://deltarho.org/docs-datadr/}
}
\examples{
help(package = datadr)
}
\author{
Ryan Hafen
Maintainer: Ryan Hafen <rhafen@gmail.com>
}
\keyword{package}

