% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ddf}
\alias{ddf}
\title{Instantiate a Distributed Data Frame ('ddf')
Instantiate a distributed data frame ('ddf')}
\usage{
ddf(conn, transFn = NULL, update = FALSE, reset = FALSE, control = NULL,
  verbose = TRUE)
}
\arguments{
\item{conn}{an object pointing to where data is or will be stored for the 'ddf' object - can be a 'kvConnection' object created from \code{\link{localDiskConn}} or \code{\link{hdfsConn}}, or a data frame or list of key-value pairs}

\item{transFn}{transFn a function to be applied to the key-value pairs of this data prior to doing any processing, that transform the data into a data frame if it is not stored as such}

\item{update}{should the attributes of this object be updated?  See \code{\link{updateAttributes}} for more details.}

\item{reset}{should all persistent metadata about this object be removed and the object created from scratch?  This setting does not effect data stored in the connection location.}

\item{control}{parameters specifying how the backend should handle things if attributes are updated (most-likely parameters to \code{rhwatch} in RHIPE) - see \code{\link{rhipeControl}} and \code{\link{localDiskControl}}}

\item{verbose}{logical - print messages about what is being done}
}
\description{
Instantiate a Distributed Data Frame ('ddf')
Instantiate a distributed data frame ('ddf')
}

