% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddo_ddf_methods.R
\name{ddo-ddf-accessors}
\alias{bsvInfo}
\alias{counters}
\alias{ddo-ddf-accessors}
\alias{getKeys}
\alias{hasExtractableKV}
\alias{kvExample}
\alias{length.ddo}
\alias{splitSizeDistn}
\title{Accessor Functions}
\usage{
kvExample(x)

bsvInfo(x)

counters(x)

splitSizeDistn(x)

getKeys(x)

hasExtractableKV(x)

\method{length}{ddo}(x)
}
\arguments{
\item{x}{a 'ddf'/'ddo' object}
}
\description{
Accessor functions for attributes of ddo/ddf objects.  Methods also include \code{nrow} and \code{ncol} for ddf objects.
}

