% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ddo_ddf_methods.R
\name{kvExample}
\alias{bsvInfo}
\alias{counters}
\alias{getKeys}
\alias{hasExtractableKV}
\alias{kvExample}
\alias{length.ddo}
\alias{names.ddf}
\alias{splitRowDistn}
\alias{splitSizeDistn}
\alias{summary.ddf}
\alias{summary.ddo}
\title{Accessor Functions}
\usage{
kvExample(x)

bsvInfo(x)

counters(x)

splitSizeDistn(x)

splitRowDistn(x)

getKeys(x)

\method{summary}{ddo}(object, ...)

\method{summary}{ddf}(object, ...)

hasExtractableKV(x)

\method{names}{ddf}(x)

\method{length}{ddo}(x)
}
\arguments{
\item{x}{a 'ddf'/'ddo' object}

\item{object}{a 'ddf'/'ddo' object}

\item{\ldots}{additional arguments}
}
\description{
Accessor functions for attributes of ddo/ddf objects.  Methods also include \code{nrow} and \code{ncol} for ddf objects.

Accessor methods for 'ddo' and 'ddf' objects
}

