% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recombine_transforms.R
\name{drGLM}
\alias{drGLM}
\title{GLM Transformation Method}
\usage{
drGLM(...)
}
\arguments{
\item{\ldots}{arguments you would pass to the \code{\link{glm}} function}
}
\description{
GLM transformation method
}
\details{
This provides a transformation function to be called for each subset in a recombination MapReduce job that applies R's glm method and outputs the coefficients in a way that \code{\link{combMeanCoef}} knows how to deal with.  It can be applied to a ddf with \code{\link{addTransform}} prior to calling \code{\link{recombine}}.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{divide}}, \code{\link{recombine}}, \code{\link{rrDiv}}
}

