% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataops_persist.R
\name{drPersist}
\alias{drPersist}
\title{Persist a Transformed 'ddo' or 'ddf' Object}
\usage{
drPersist(x, output = NULL, overwrite = FALSE, control = NULL)
}
\arguments{
\item{x}{an object of class 'ddo' or 'ddf'}

\item{output}{a "kvConnection" object indicating where the output data should reside (see \code{\link{localDiskConn}}, \code{\link{hdfsConn}}).  If \code{NULL} (default), output will be an in-memory "ddo" object.}

\item{overwrite}{logical; should existing output location be overwritten? (also can specify \code{overwrite = "backup"} to move the existing output to _bak)}

\item{control}{parameters specifying how the backend should handle things (most-likely parameters to \code{rhwatch} in RHIPE) - see \code{\link{rhipeControl}} and \code{\link{localDiskControl}}}
}
\value{
a 'ddo' or 'ddf' object with the transformation evaluated on the data
}
\description{
Persist a transformed 'ddo' or 'ddf' object by making a deferred transformation permanent
}
\details{
When a transformation is added to a ddf/ddo via \code{\link{addTransform}}, the transformation is deferred until
the some action is taken with the data (e.g. a call to \code{\link{recombine}}).  See the documentation of
\code{\link{addTransform}} for more information about the nature of transformations.

Calling \code{drPersist()} on the ddo/ddf makes the transformation permanent (persisted).  In the case of a local disk
connection (via \code{\link{localDiskConn}}) or HDFS connection (via \code{\link{hdfsConn}}), the transformed data
are written to disc.
}
\examples{
bySpecies <- divide(iris, by = "Species")

# Create the transformation and add it to bySpecies 
bySpeciesSepal <- addTransform(bySpecies, function(x) x[,c("Sepal.Length", "Sepal.Width")])

# Note the transformation is 'pending' a data action 
bySpeciesSepal

# Make the tranformation permanent (persistent)
bySpeciesSepalPersisted <- drPersist(bySpeciesSepal)

# The transformation no longer pending--but a permanent part of the new ddo
bySpeciesSepalPersisted
bySpeciesSepalPersisted[[1]]
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{addTransform}}
}

