% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divSpec_condDiv.R
\name{getCondCuts}
\alias{getCondCuts}
\title{Get names of the conditioning variable cuts}
\usage{
getCondCuts(df, splitVars)
}
\arguments{
\item{df}{a data frame}

\item{splitVars}{a vector of variable names to split by}
}
\description{
Used internally for exported for certain reasons.  Do not use explicitly.
}

