% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/divide.R
\name{getSplitVars}
\alias{getSplitVars}
\title{Extract "Split" Variables}
\usage{
getSplitVars(x)
}
\arguments{
\item{x}{a key-value pair or a value}
}
\description{
For a given k/v pair or value, exract all split variables (split variables are variables that define how the data was divided).
}

