% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kvPairs}
\alias{kvPairs}
\title{Manually Specify Key-Value Pairs}
\usage{
kvPairs(...)
}
\arguments{
\item{\ldots}{key-value pairs (lists with two elements)}
}
\value{
a list of objects of class "kvPair"
}
\description{
Manually specify key-value pairs
}
\author{
Ryan Hafen
}

