% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvPairs.R
\name{kvPairs}
\alias{kvPairs}
\title{Specify a Collection of Key-Value Pairs}
\usage{
kvPairs(...)
}
\arguments{
\item{\ldots}{key-value pairs (lists with two elements)}
}
\value{
a list of objects of class "kvPair"
}
\description{
Specify a collection of key-value pairs
}
\examples{
kvPairs(kvPair(1, letters), kvPair(2, rnorm(10)))
}
\seealso{
\code{\link{kvPair}}
}

