% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeExtractable}
\alias{makeExtractable}
\title{Take a ddo/ddf HDFS data object and turn it into a mapfile}
\usage{
makeExtractable(obj)
}
\arguments{
\item{obj}{object of class 'ddo' or 'ddf' with an HDFS connection}
}
\description{
Take a ddo/ddf HDFS data object and turn it into a mapfile
}

