% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddo_ddf_kvHDFS.R
\name{makeExtractable}
\alias{makeExtractable}
\title{Take a ddo/ddf HDFS data object and turn it into a mapfile}
\usage{
makeExtractable(obj)
}
\arguments{
\item{obj}{object of class 'ddo' or 'ddf' with an HDFS connection}
}
\description{
Take a ddo/ddf HDFS data object and turn it into a mapfile
}

