% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.ddo}
\alias{print.ddo}
\title{Print a "ddo" or "ddf" Object}
\usage{
\method{print}{ddo}(x, ...)
}
\arguments{
\item{x}{object to be printed}

\item{\ldots}{additional arguments}
}
\description{
Print an overview of attributes of distributed data objects (ddo) or distributed data frames (ddf)
}
\author{
Ryan Hafen
}

