% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddo_ddf_methods.R
\name{removeData}
\alias{removeData}
\title{Remove Key-Value Pairs from a Data Connection}
\usage{
removeData(conn, keys)
}
\arguments{
\item{conn}{a kvConnection object}

\item{keys}{a list of keys indicating which k/v pairs to remove}
}
\description{
Remove key-value pairs from a data connection
}
\note{
This is generally not recommended for HDFS as it writes a new file each time it is called, and can result in more individual files than Hadoop likes to deal with.
}
\author{
Ryan Hafen
}
\seealso{
\code{\link{removeData}}, \code{\link{localDiskConn}}, \code{\link{hdfsConn}}
}

