% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddo_addTransform.R
\name{setupTransformEnv}
\alias{setupTransformEnv}
\title{Set up transformation environment}
\usage{
setupTransformEnv(transFns, env = NULL)
}
\arguments{
\item{transFns}{from the "transforms" attribute of a ddo object}

\item{env}{the environment in which to evaluate the transformations}
}
\description{
This is called internally in the map phase of datadr MapReduce jobs.  It is not meant for use outside of there, but is exported for convenience.
Given an environment and collection of transformations, it populates the environment with the global variables in the transformations.
}

